ADMIN 검증 메서드 추가,
ADMIN.PW application.properties 에 추가

+++ src/main/java/LD_Caffe/ld_caffe/controller/AdminController.java
@@ -1,14 +1,15 @@
 package LD_Caffe.ld_caffe.controller;
 
+import LD_Caffe.ld_caffe.dto.LoginDto;
 import LD_Caffe.ld_caffe.dto.MenuDto;
-import LD_Caffe.ld_caffe.repository.MenuRepository;
 import LD_Caffe.ld_caffe.service.AdminService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
+import org.springframework.security.core.Authentication;
 import org.springframework.util.StringUtils;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
@@ -17,7 +18,6 @@ import java.io.File;
 import java.io.IOException;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
 
@@ -28,25 +28,20 @@ import java.util.UUID;
 public class AdminController {
 
     private final AdminService adminService;
+    @Value("${ADMIN.PW}")
+    private String ADMIN_PW;
 
 
-    @ResponseBody
-    @GetMapping("/users")
-    public ArrayList<String> AlluserInfo(){
-
-//        로그인한 사람이 관리자 계정인지 확인하기
-
-//        ArrayList<String> usersInfo = new ArrayList<>();
-//        usersInfo.add("TestData1");
-//        usersInfo.add("TestData2");
-//        usersInfo.add("TestData3");
-
-        return adminService.getUserNames();
+    private boolean amIAdmin(Authentication authentication){  // ADMIN 계정 확인 메서드
+        return authentication.getName().equals("ADMIN"); // ADMIN 이 맞다면 TRUE 아니면 FALSE
     }
 
-    @DeleteMapping("/users/delete/{name}")
-    public ResponseEntity<String> deleteUserInfo(@PathVariable String name){
-        System.out.println(name);
+    @DeleteMapping("/users/delete/{name}")  // 유저 삭제 메서드
+    public ResponseEntity<String> deleteUserInfo(@PathVariable String name,Authentication authentication){
+
+        if (!amIAdmin(authentication)) { // ADMIN 이 아니라면 여기서 리턴문으로 빠져나감
+            return ResponseEntity.status(403).body("유효한 접근이 아닙니다.");
+        }
 
         if (adminService.deleteUser(name)){
             System.out.println("데이터 찾음 // 데이터 삭제 완료");
@@ -55,16 +50,20 @@ public class AdminController {
             System.out.println("데이터 없음 // 데이터 삭제 실패");
             return ResponseEntity.notFound().build();
         }
-
     }
 
-    // 메뉴 데이터베이스 저장
+    // 메뉴 데이터베이스 저장 메서드
     @PostMapping("/menu")
     public ResponseEntity<String> addMenu(@RequestParam("image") MultipartFile file,
                                           @RequestParam("category") String category,
                                           @RequestParam("name") String name,
                                           @RequestParam("content") String content,
-                                          @RequestParam("price") Integer price){
+                                          @RequestParam("price") Integer price,
+                                          Authentication authentication){
+
+        if (!amIAdmin(authentication)) {  // 어드민이 아니라면 여기서 리턴문으로 블락
+            return ResponseEntity.status(403).body("유효한 접근이 아닙니다.");
+        }
 
         System.out.println("<< 추가할 메뉴 >>");
         System.out.println("Category : " +category);
@@ -78,7 +77,7 @@ public class AdminController {
             String fileName = UUID.randomUUID().toString() + "_" + StringUtils.cleanPath(file.getOriginalFilename());
 
             // 이미지 파일 저장 경로
-            String saveDir = "/Users/jujaeyoung/desktop/images/";
+            String saveDir = "C:/VSCODE workspace/LD_Front/Front/public/img";
             Path filePath = Paths.get(saveDir, fileName);
 
             // 이미지 파일 저장
@@ -105,7 +104,7 @@ public class AdminController {
         }
     }
 
-    @GetMapping("/menuList")
+    @GetMapping("/menuList")  // 메뉴 리스트 메서드
     public ResponseEntity<List<MenuDto>> MenuList() throws IOException {
 
         List<MenuDto> menuInfo = adminService.getAllMenuInfo();
@@ -120,10 +119,12 @@ public class AdminController {
         return new ResponseEntity<>(menuInfo, headers, HttpStatus.OK);
     }
 
-    @DeleteMapping("/menu/delete/{menuName}")
-    public ResponseEntity<String> deleteMenuInfo(@PathVariable String menuName){
+    @DeleteMapping("/menu/delete/{menuName}")   // 메뉴삭제 메서드
+    public ResponseEntity<String> deleteMenuInfo(@PathVariable String menuName,Authentication authentication){
+        if (!amIAdmin(authentication)) { // ADMIN 이 아니라면 여기서 블락
+            return ResponseEntity.status(403).body("유효한 접근이 아닙니다.");
+        }
         System.out.println(menuName);
-
         try{
             if (adminService.deleteMenu(menuName)){
                 System.out.println(" 메뉴 삭제 완료");
@@ -139,6 +140,24 @@ public class AdminController {
         }
     }
 
+    @PostMapping("/login")
+    public ResponseEntity<String> adminLogin(@RequestBody LoginDto loginDto){ // 어드민 로그인, 토큰발행 메서드
+        System.out.println("ADMIN_PW = " + ADMIN_PW);
+        System.out.println("어드민 비밀번호 시도 = " + loginDto.getU_pw());
+        if (loginDto.getU_pw().equals(ADMIN_PW)) { // 유저가 입력한 비밀번호가 어드민 비밀번호가 맞는경우
+            LoginDto admin = new LoginDto();
+            admin.setU_id("ADMIN");
+            admin.setU_pw(ADMIN_PW);
+            String token = adminService.createAdminToken(admin);  // ADMIN 이라는 ID 값을 가진 토큰 발행
+            if(token.equals("0")){  // 로그인할때 쓰던 코드 복붙
+                return ResponseEntity.badRequest().build();
+            }else{
+                System.out.println("ADMIN_token = " + token);
+                return ResponseEntity.ok().body(token);
+            }
+        }else{
+            return ResponseEntity.badRequest().body("올바르지 않은 비밀번호입니다.");
+        }
 
-
+    }
 }
